requirements.txt
django
python
django-phonenumber-field
phonenumbers
djangorestframework
django-allauth
django-rest-auth
pillow